introduction proofread
Just a couple of small questions...
p. 27: Is "random discrepancies" appropriate here, or should something
like "occasional discrepancies" be used? I couldn't tell if the
repetition of the word "random" in this case was intentional.
p. 32: I decided that the parameters looked better as just ital. rather
than just bold—OK?
p. 38: I would like to see "Perlin himself" and "noise detail ()
function" together instead of separated by the URL.  I tried to fix
this in the first instance, but not sure if I did it correctly!